+++ AgeCalculator/nbproject/project.properties
@@ -53,7 +53,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=cst1201.AgeCalculator
+main.class=cst1201.Age_Calculator
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
